LayoutTests/fullscreen/full-screen-iframe-legacy-expected.txt

Test for bug 61461: Handle entering full screen security restrictions

To test manually, click the video "full screen" button - the page should enter full screen mode.

EVENT(webkitfullscreenchange)
TEST(document.getElementById('frame').contentDocument.width==document.width) OK
END OF TEST
